Reshuffle headings for easier comprehension
As it stands, the current documentation does not distinguish between the
settings.yml file and the custom parameters on join. These are
fundamentally different in nature, so it is useful to point this out via
separate headings. A few words have been tweaked for general comprehension.
The hyperlink to the tool has also been added so the user can directly
click on the link.