Complete documentation for userdata-* params
Custom parameters are an extremely useful addition to our BBB toolset.
It is a pity that these have gone undocumented for a while now and this
commit fixes that. It separates and categorizes userdata-* params
according to the addUserSettings.js file and presents them in a clear
and concise table format. Descriptions are specified and default values
are added. An example is also included for reference.