Add new method of installation for larger deployments
Increasingly, more people want to deploy BigBlueButton at a larger scale.
The existing methods are not suitable for scaling to, say, 10 different
servers. This commit adds guidelines for larger deployments including
several community-sourced Ansible roles for inspiration. This is an
immensely useful resource for developers and maintainers of larger
systems.