Decouple files to ensure greater clarity
The process of getting help is different to that of actually
troubleshooting. This commit makes it much easier for a user
to seek help by first navigating to the getting started page
and then choosing which route they decide to follow. It is
expected that once the demand for BBB stabilizes in the future
we can move getting help lower down, perhaps.